requirements.txt
pyserial==3.5
inputs==0.5